Verschiedene kleinere Fehler die beim Belastungs-Test auftraten behoben. Python-Script um die Ergebnisse vom JMeter zu verarbeiten. HBase-Einstellungen für den privaten HBase und Hadoop-Cluster (immernoch mit zwei ZooKeepern). User Service erlaubt jetzt beliebige gleichzeitige Logins. Sessions werden grundsätzlich nach 24h gelöscht - falls kein expliziter logout erfolgt.
git-svn-id: svn://localhost/twospot_svn/trunk@1109 8d73be03-56cb-46cc-8948-98b08ae92ef2

+++ TwoSpotJdo/src/org/prot/jdo/storage/StoragePersistenceHandler.java
@@ -79,10 +79,10 @@ public class StoragePersistenceHandler implements StorePersistenceHandler
 
 			// Get the primary key
 			Object pKeyObj = sm.provideField(sm.getClassMetaData().getPKMemberPositions()[0]);
-			assert (pKeyObj instanceof Key);
 			Key key = (Key) pKeyObj;
+			if (key == null)
+				return;
 
-			// Delete the object
 			Storage storage = mconn.getStorage();
 			storage.deleteObject(appId, kind, key);
